refactor : style 코드 분리
